pyproject.toml
[project]
name = "econflux"
version = "0.1.0"
description = "Agentic financial intelligence system powered by Strands-Agents, yfinance, and Bedrock AgentCore."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=9.0.1",
    "strands-agents>=1.18.0",
    "bedrock-agentcore>=1.0.6",
    "yfinance>=0.2.66",
    "strands-agents-builder>=0.1.10",
    "strands-agents-tools>=0.2.16",
]
